// packages/receipt/src/generated/types.ts
export interface BasicHeaderType {
  /**
   * Identifier of the request/response, unique with the taxnumber in every data
   * exchange transaction
   *
   * A kérés/válasz azonosítója, minden üzenetváltásnál - adószámonként - egyedi
   */
  requestId: string;
  /**
   * UTC time of the request/response
   *
   * A kérés/válasz keletkezésének UTC ideje
   */
  timestamp: string;
  /**
   * Request version number, indicating which datastructure the client sends data
   * in, and in which the response is expected
   *
   * A kérés/válasz verziószáma, hogy a hívó melyik interfész verzió szerint küld
   * adatot és várja a választ
   */
  requestVersion: string;
  /**
   * Header version number
   *
   * A header verziószáma
   */
  headerVersion?: string;
}
// ...
export interface BasicResultType {
  /**
   * Processing result
   *
   * Feldolgozási eredmény
   */
  funcCode: FunctionCodeType;
  /**
   * Processing error code
   *
   * A feldolgozási hibakód
   */
  errorCode?: string;
  /**
   * Processing message
   *
   * Feldolgozási üzenet
   */
  message?: string;
  /**
   * Miscellaneous notifications
   *
   * Egyéb értesítések
   */
  notifications?: NotificationsType;
}
// ...
export interface CashRegisterEventType {
  /**
   * Event code
   *
   * Esemény kódja
   */
  eventCode: EventCodeType;
  /**
   * Event descriptive data
   *
   * Esemény leíró adata
   */
  eventDescription?: string;
}
// ...
export interface CashRegisterInfoResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
}
// ...
export interface CashRegisterInfoType {
  /**
   * Event list of the e-cash register. Events occured after the previous "Cash
   * Register Info" report should be sent.
   *
   * Az e-pénztárgépben bekövetkezett események naplója. Az utolsó
   * adatszolgáltatás óta bekövetkezett eseményeket kell beküldeni.
   *
   * @remarks At most 1000 entries.
   */
  cashRegisterEvent?: CashRegisterEventType[];
  /**
   * Technical parameters of the e-cash register according to the current state
   * of the e-cash register.
   *
   * Az e-pénztárgép műszaki jellemzői, az e-pénztárgép aktuális állapotának
   * megfelelően.
   */
  cashRegisterTechnicalInfo: CashRegisterTechnicalInfoType;
  /**
   * System state statistics of the e-cash register, according to the current
   * state of the e-cash register. In case of errors: All errors occured after
   * the previous "Cash Register Info" report should be sent.
   *
   * e-Pénztárgép rendszerállapot statisztika, az e-pénztárgép aktuális
   * állapotának megfelelően, illetve a hibák tekintetében az utolsó
   * adatszolgáltatás óta.
   */
  cashRegisterState: CashRegisterStateType;
  /**
   * Geographical location of the e-cash register. Only applicable if the e-cash
   * register has a GPS unit.
   *
   * Az e-pénztárgép földrajzi helye. Csak a műholdas helymeghatározásra alkalmas
   * e-pénztárgép esetén kell kitölteni az adatot.
   */
  cashRegisterPosition?: CashRegisterPositionType;
  /**
   * In case the clock of the e-cash register has been set, then data of the
   * clock setting.
   *
   * Amennyiben az e-pénztárgép órája állításra került, akkor az óraállítás
   * körülményeiről adatot kell szolgáltatni.
   */
  cashRegisterTimeUpdate?: CashRegisterTimeUpdateType;
}
// ...
export interface CashRegisterPositionType {
  /**
   * Latitude of e-cash register in degrees, maximum to 5 decimal places, using a
   * sign, in WGS84 standard
   *
   * e-Pénztárgép földrajzi szélessége fokban, legfeljebb 5 tizedesjegy
   * pontossággal, előjelet használva, a WGS84 szabvány szerint
   */
  latitude: string;
  /**
   * Longitude of e-cash register in degrees, maximum to 5 decimal places, using
   * a sign, in WGS84 standard
   *
   * e-Pénztárgép földrajzi hosszúsága fokban, legfeljebb 5 tizedesjegy
   * pontossággal, előjelet használva, a WGS84 szabvány szerint
   */
  longitude: string;
  /**
   * e-Cash register altitude above sea level in metres, according to WGS84
   *
   * e-Pénztárgép tengerszint feletti magassága méterben, a WGS84 szabvány
   * szerint
   */
  height: number;
}
// ...
export interface CashRegisterStateType {
  /**
   * e-Cash register work state
   *
   * e-Pénztárgép állapot
   */
  cashRegisterWorkState: CashRegisterWorkStateType;
  /**
   * AE block state
   *
   * AE blokkolási állapot
   */
  aeBlockUnblockState: AeBlockUnblockStateType;
  /**
   * Fiscal day state
   *
   * Adóügyi nap állapota
   */
  fiscalDayState: FiscalDayStateType;
  /**
   * The recordCounter value of the last sent document. See further information
   * in the "Generate NAV verification code" chapter of the Development
   * Documentation.
   *
   * Az utoljára beküldött bizonylat recordCounter értéke. Leírását lásd a
   * Fejlesztői dokumentáció „NAV ellenőrző kód képzése” fejezetében.
   */
  lastSentRecordCounter: number;
  /**
   * The recordCounter value of the last document in the e-cash register. See
   * further information in the "Generate NAV verification code" chapter of the
   * Development Documentation.
   *
   * Az e-pénztárgépen utoljára kiállított bizonylat recordCounter értéke.
   * Leírását lásd a Fejlesztői dokumentáció „NAV ellenőrző kód képzése”
   * fejezetében.
   */
  lastRecordCounter: number;
  /**
   * AE state. It should be filled only by HW based e-cash register.
   *
   * AE állapot. Csak HW alapú e-pénztárgép esetén kell kitölteni.
   */
  aeState?: AeStateType;
  /**
   * List of internal error codes that have occurred in the e-cash register since
   * the last cashRegisterState record was sent
   *
   * e-Pénztárgépben az utolsó cashRegisterState rekord küldése óta jelentkezett
   * hibák belső kódjainak listája
   */
  errors?: string;
}

/**
 * e-Cash register technical features
 *
 * e-Pénztárgép műszaki jellemzői
 */
export interface CashRegisterTechnicalInfoType {
  /**
   * Details of the software on the e-cash register
   *
   * Az e-pénztárgépen futó szoftver adatai
   */
  cashRegisterSoftware: SoftwareType;
  /**
   * The IMEI number of the modem/cash register in the AE
   *
   * Az AE-ben lévő modem/ePénztárgép kiolvasott IMEI száma.
   */
  imei: string;
  /**
   * The IMSI number of the SIM in the AE/cash register
   *
   * Az AE-ben/ePénztárgépben lévő SIM kiolvasott IMSI száma.
   */
  imsi: string;
  /**
   * The operator or site data of the e-cash register.
   *
   * Az e-pénztárgép üzemeltetői és üzemeltetési hely adatai.
   */
  operatorSite: OperatorSiteType;
  /**
   * Current VAT table
   *
   * Aktuális ÁFA-tábla (a vonatkozó NAV üzenet)
   */
  vat: VatType;
  /**
   * Latest used installation code
   *
   * Legutóbb használt üzembehelyezési kód
   */
  registrationNumber: string;
  /**
   * The version number of the communication protocol used by the e-cash register
   *
   * Az e-pénztárgép által használt kommunikációs protokoll verziószáma
   */
  communicationProtocolVersion: string;
  /**
   * Latest schema version number
   *
   * A legfrissebb séma verziószáma, aminek az e-pénztárgép által képzett
   * üzenetek megfelelnek.
   */
  schemaVersion: string;
  /**
   * EPD process IDs of the executed instructions that produce the current state
   * of the e-cash register, and the last received but not yet executed (timed)
   * instructions
   *
   * Az e-pénztárgép jelenlegi állapotát előállító végrehajtott utasítások,
   * valamint a legutolsó átvett, de még nem végrehajtott (időzített) utasítások
   * EPD folyamat azonosítói
   */
  epdProcessIdentifiers: EpdProcessIdentifiersType;
  /**
   * HW-based e-cash register technical features. It should be filled only by HW
   * based e-cash register.
   *
   * HW alapú e-pénztárgép műszaki jellemzői. Csak HW alapú e-pénztárgép esetén
   * kell kitölteni.
   */
  hardwareTechnicalInfo?: HardwareTechnicalInfoType;
  /**
   * Bank account number assigned to the POS terminal
   *
   * A POS terminálhoz rendelt bankszámla száma
   */
  posAssignedBankAccounts?: PosAssignedBankAccountListType;
}
// ...
export type DocumentClassType =
  | 'RECEIPT'
  | 'FUEL'
  | 'HEALTH'
  | 'DEFERRED'
  | 'CONSUME'
  | 'SIMPLIFIEDINVOICE'
  | 'NORMALINVOICE'
  | 'OTHER';
// ...
export interface DocumentResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * The tax number in the request.
   *
   * A kérésben szereplő adószám.
   */
  taxNumber: string;
}
// ...
export interface GetProductByCodeResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * The number of products in the results list.
   *
   * Az eredménylistában található termékek száma.
   */
  numberOfProducts: number;
  /**
   * A results list containing products that match the search criteria.
   *
   * A keresési feltételnek megfelelt termékeket tartalmazó eredménylista.
   */
  products: ProductListType;
}
// ...
export interface HelloResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
}
// ...
export interface ProductListType {
  /**
   * Type of result list containing products that match the search criteria.
   *
   * A keresési feltételnek megfelelt termékeket tartalmazó eredménylista típusa.
   *
   * @remarks At most 100 entries.
   */
  productItem: ProductItemType[];
}
// ...
export interface QueryTaxpayerResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * Last date on which the data was changed
   *
   * Az adatok utolsó változásának időpontja
   */
  infoDate?: string;
  /**
   * Indicates whether the queried taxpayer is existing and valid
   *
   * Jelzi, hogy a lekérdezett adózó létezik és érvényes-e
   */
  taxpayerValidity: boolean;
  /**
   * Response data of the taxpayer query
   *
   * Az adózó lekérdezés válasz adatai
   */
  taxpayerData?: TaxpayerDataType;
}
// ...
export type ReportClassType =
  'CASHREGISTEROPENBALANCE' | 'DAILYCASHFLOW' | 'CASHREGISTER' | 'CASHFLOW' | 'OTHER';
// ...
export interface ReportResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * The tax number in the request.
   *
   * A kérésben szereplő adószám.
   */
  taxNumber: string;
}